requirements.txt
eth-brownie
pandas
pymongo
tensorflow
numpy
gym
